+++ test/polyfills/EventTarget.mjs
@@ -2,4 +2,4 @@
 
 import { EventTarget } from "event-target-shim";
 
-if (!("EventTarget" in global)) global.EventTarget = EventTarget;
+if (!("EventTarget" in globalThis)) globalThis.EventTarget = EventTarget;