+++ changelog.md
@@ -21,6 +21,7 @@
 - Removed the now redundant `not IE > 0` from the Browserslist query.
 - Updated `react-dom/server` imports in tests to suit React v18.
 - Fixed the `fetchGraphQL` test with the global `fetch` API unavailable for new versions of Node.js that have the `fetch` global.
+- Use `globalThis` instead of `global` in tests.
 - Fixed some JSDoc links.
 - Removed some unnecessary JSDoc comments in tests.
 - Revamped the readme: